backend/requirements.txt
cohere
qdrant-client
beautifulsoup4
requests
python-dotenv
fastapi
openai
uvicorn
httpx
google.generativeai